requirements.txt
streamlit
pandas
Pillow
PyMuPDF
